+++ src/Restorer.php
@@ -43,15 +43,15 @@ final class Restorer
         }
     }
 
-    public function restoreStaticAttributes(Snapshot $snapshot): void
+    public function restoreStaticProperties(Snapshot $snapshot): void
     {
         $current    = new Snapshot($snapshot->excludeList(), false, false, false, false, true, false, false, false, false);
         $newClasses = array_diff($current->classes(), $snapshot->classes());
 
         unset($current);
 
-        foreach ($snapshot->staticAttributes() as $className => $staticAttributes) {
-            foreach ($staticAttributes as $name => $value) {
+        foreach ($snapshot->staticProperties() as $className => $staticProperties) {
+            foreach ($staticProperties as $name => $value) {
                 $reflector = new ReflectionProperty($className, $name);
                 $reflector->setAccessible(true);
                 $reflector->setValue($value);
@@ -62,14 +62,14 @@ final class Restorer
             $class    = new ReflectionClass($className);
             $defaults = $class->getDefaultProperties();
 
-            foreach ($class->getProperties() as $attribute) {
-                if (!$attribute->isStatic()) {
+            foreach ($class->getProperties() as $property) {
+                if (!$property->isStatic()) {
                     continue;
                 }
 
-                $name = $attribute->getName();
+                $name = $property->getName();
 
-                if ($snapshot->excludeList()->isStaticAttributeExcluded($className, $name)) {
+                if ($snapshot->excludeList()->isStaticPropertyExcluded($className, $name)) {
                     continue;
                 }
 
@@ -77,8 +77,8 @@ final class Restorer
                     continue;
                 }
 
-                $attribute->setAccessible(true);
-                $attribute->setValue($defaults[$name]);
+                $property->setAccessible(true);
+                $property->setValue($defaults[$name]);
             }
         }
     }